shared/pyproject.toml
[tool.poetry]
name = "geofoncier-shared"
version = "0.1.0"
description = "Geofoncier – shared utilities across services"
authors = []
packages = [{ include = "geofoncier_shared" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
redis = { version = ">=5.0", extras = ["hiredis"] }
pydantic-settings = ">=2.0"
fastapi = ">=0.111.0"
PyJWT = { version = ">=2.8", extras = ["crypto"] }

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
pytest-cov = ">=5.0"
cryptography = ">=42.0"
black = ">=24.0"
httpx = ">=0.27"
pytest-mock = ">=3.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["geofoncier_shared"]
omit = ["geofoncier_shared/__init__.py"]

[tool.black]
target-version = ["py312"]
